add render

+++ run_holodeck.py
@@ -9,11 +9,18 @@ args_path = f"args_{method}.json"
 roominfo_path = f"roominfo_{method}.json"
 
 # roomtype = "bedroom"
-for  roomtype in ["bedroom","livingroom"]:
-    for idx in range(10):
+basedir = "/mnt/fillipo/yandan/scenesage/Holodeck0509/"
+outdir = "/mnt/fillipo/yandan/scenesage/record_scene/holodeck/"
+roomtypes = os.listdir(basedir)
+for  roomtype in roomtypes:
+    if roomtype not in ["bathroom","children_room","classroom","gameroom","gym","kitchen","laundromat","meetingroom","office","restaurant","waitingroom"]:
+        continue
+    os.makedirs(f"{outdir}/{roomtype}/",exist_ok=True)
+    for idx in range(3):
         # for idx in [11,13,18]:
         print(f"\n=== Running Task {idx} ===")
-        jsonname = f"/mnt/fillipo/yandan/scenesage/Holodeck/{roomtype}/holodeck_{roomtype}_{idx}.json"
+        jsonname = f"{basedir}/{roomtype}/{roomtype}_{idx}.json"
+        print(f"!!!!processing {jsonname}")
         # Edit roominfo_idesign.json
         with open(roominfo_path, "r") as f:
             roominfo_data = json.load(f)
@@ -26,8 +33,10 @@ for  roomtype in ["bedroom","livingroom"]:
         room_size = [i + 0.28 for i in room_size]
         roominfo_data["roomsize"] = room_size
         roominfo_data["save_dir"] = (
-            f"/mnt/fillipo/yandan/scenesage/record_scene/{method}/{roomtype}_{idx}"
+            f"{outdir}/{roomtype}/{roomtype}_{idx}"
         )
+        if os.path.exists(roominfo_data["save_dir"]):
+            continue
         with open(roominfo_path, "w") as f:
             json.dump(roominfo_data, f, indent=4)
