+++ run_layoutgpt.py
@@ -7,18 +7,27 @@ import time
 args_path = "args_layoutgpt.json"
 roominfo_path = "roominfo_layoutgpt.json"
 
+basedir = "/mnt/fillipo/yandan/scenesage/LayoutGPT0509/"
+outdir = "/mnt/fillipo/yandan/scenesage/record_scene/layoutgpt/"
 # roomtype = "livingroom"
-for  roomtype in ["bedroom","livingroom"]:
-    for i in range(10):
+roomtypes = os.listdir(basedir)
+for  roomtype in roomtypes: # ["bedroom","livingroom"]:
+    if roomtype not in ["bathroom","children_room","classroom","gameroom","gym","kitchen","laundromat","meetingroom","office","restaurant","waitingroom"]:
+        continue
+    os.makedirs(f"{outdir}/{roomtype}/",exist_ok=True)
+    for i in range(1,4):
         # for i in [11,13,18]:
-        print(f"\n=== Running Task {i} ===")
+        
 
         # Edit args_idesign.json
         with open(args_path, "r") as f:
             args_data = json.load(f)
-        args_data["json_name"] = (
-            f"/mnt/fillipo/yandan/scenesage/LayoutGPT/{roomtype}_scene_graph/layoutgpt_{roomtype}_{i}.json"
-        )
+        # args_data["json_name"] = (
+        #     f"/mnt/fillipo/yandan/scenesage/LayoutGPT/{roomtype}_scene_graph/layoutgpt_{roomtype}_{i}.json"
+        # )
+        args_data["json_name"] = f"{basedir}/{roomtype}/{roomtype}_%02d.json"  % i
+        json_name = args_data["json_name"]
+        print(f"\n=== Running Task {json_name}===")
         with open(args_path, "w") as f:
             json.dump(args_data, f, indent=4)
 
@@ -28,14 +37,17 @@ for  roomtype in ["bedroom","livingroom"]:
 
         with open(args_data["json_name"], "r") as f:
             j = json.load(f)
-            roomsize = j["room_size"]
+            # roomsize = j["room_size"]
+            roomsize = j["roomsize"]
 
         room_size = [roomsize["width"], roomsize["length"]]
         room_size = [i + 0.28 for i in room_size]
         roominfo_data["roomsize"] = room_size
         roominfo_data["save_dir"] = (
-            f"/mnt/fillipo/yandan/scenesage/record_scene/layoutgpt/{roomtype}_{i}"
+            f"{outdir}/{roomtype}/{roomtype}_{i}"
         )
+        if os.path.exists(roominfo_data["save_dir"]):
+            continue
         with open(roominfo_path, "w") as f:
             json.dump(roominfo_data, f, indent=4)
